Check `de/docs/fastapi-cli.md`
Aside of appending to the de llm-prompt, formulate a sentence in the translation script's general prompt more precise, to prevent that it translates text nodes in html in codeblocks (like in the current `de/docs/fastapi-cli.md`). That it translates comments (like in the current `de/docs/environment-variables.md`) is useful, so explicitely allow this.

If this does not work, an idea may be to replace the codeblocks with something like `donotchangeme` before handing the text over to the AI, inserting them back afterwards.