Strip codeblocks from the text before translation
... as the AIs seem to have too big problems understanding what a codeblock is.

For this we add two functions `extract_codeblocks` and `add_back_codeblocks`, which do the work. And we modify the `translate_page` function in a few places to make use of the two new functions.

A positive side effect of this is that there is less to translate for the AIs, so less costs.