Update #13930
This time I regenerated the translations, using the existing translations from #13930, and some wordlists, not from scratch, as the previous commit. There was one case where the AI did not remove a paragraph, which was removed from the english text, see query-params-str-validations.md, but otherwise this went smooth. Rarely were manual changes necessary. Is now in sync with the English docs.

You may apply this commit to #13930, Regeneration is not necessary.